requirements.txt
openai
transformers
torch
google-cloud-speech
pyannote.audio
python-docx
pyaudio
